Ağ arayüzünü sadece belirli koşullarda taşı
Koşul: ifupdown ile DHCP olarak yapılandırılmış ethernet arayüzleri

+++ lib/scripts/base/network.sh
@@ -4,39 +4,37 @@
 
 set -euo pipefail; [[ -z ${TRACE:-} ]] || set -x
 
-# Prefer networkd over legacy ifupdown
+# Prefer networkd over legacy ifupdown for simple cases
 
-# XXX: Overly simplistic script to migrate network configuration.
-#      Note that the migrated configuration has some bold assumptions,
-#      i.e. interface names matches "e*", and using DHCP.
+# DHCP configured ethernet interface through ifupdown?
+if [[ ! -f /etc/network/interfaces ]]; then
+	echo >&2 'No ifupdown detected; skipping'
+	exit 0
+fi
+if ! grep -Eq '^\s*iface\s+e\w+\s+inet\s+dhcp\b' /etc/network/interfaces; then
+	echo >&2 'Network interface is not suitable for migration; skipping'
+	exit 0
+fi
 
-case "$(lsb_release -sc)" in
-xenial|jessie|stretch|sid)
-	systemctl enable systemd-networkd.service
-	systemctl enable systemd-resolved.service
+systemctl enable systemd-networkd.service
+systemctl enable systemd-resolved.service
 
-	rm -f /etc/resolv.conf
-	ln -s /run/systemd/resolve/resolv.conf /etc/resolv.conf
+rm -f /etc/resolv.conf
+ln -s /run/systemd/resolve/resolv.conf /etc/resolv.conf
 
-	cat >/etc/systemd/network/99-dhcp.network <<-EOF
-		[Match]
-		Name=e*
+cat >/etc/systemd/network/99-dhcp.network <<-EOF
+	[Match]
+	Name=e*
 
-		[Network]
-		DHCP=yes
-	EOF
-	chmod a+r /etc/systemd/network/99-dhcp.network
+	[Network]
+	DHCP=yes
+EOF
+chmod a+r /etc/systemd/network/99-dhcp.network
 
-	rm -f /etc/network/interfaces
+systemctl stop networking.service
 
-	if [[ -n ${network_activate_new:-} ]]; then
-		systemctl stop networking.service
-		systemctl start systemd-networkd.service
-		systemctl start systemd-resolved.service
+systemctl start systemd-networkd.service
+systemctl start systemd-resolved.service
 
-		systemctl --no-reload disable networking.service
-	else
-		systemctl disable networking.service
-	fi
-	;;
-esac
+rm -f /etc/network/interfaces
+systemctl disable networking.service
